fix: rename packages to match go conventions
