+++ cmd/root.go
@@ -30,7 +30,7 @@ import (
 	"github.com/0xPolygon/polygon-cli/cmd/fund"
 	"github.com/0xPolygon/polygon-cli/cmd/hash"
 	"github.com/0xPolygon/polygon-cli/cmd/loadtest"
-	"github.com/0xPolygon/polygon-cli/cmd/metricsToDash"
+	"github.com/0xPolygon/polygon-cli/cmd/metricstodash"
 	"github.com/0xPolygon/polygon-cli/cmd/mnemonic"
 	"github.com/0xPolygon/polygon-cli/cmd/monitor"
 	"github.com/0xPolygon/polygon-cli/cmd/monitorv2"
@@ -142,7 +142,7 @@ func NewPolycliCommand() *cobra.Command {
 		fund.FundCmd,
 		hash.HashCmd,
 		loadtest.LoadtestCmd,
-		metricsToDash.MetricsToDashCmd,
+		metricstodash.MetricsToDashCmd,
 		mnemonic.MnemonicCmd,
 		monitor.MonitorCmd,
 		monitorv2.MonitorV2Cmd,
